requirements.txt
asgiref==3.2.7
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.5
django-appconf==1.0.4
django-braces==1.14.0
django-extra-fields==2.0.5
django-imagekit==4.0.2
django-leaflet==0.26.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
djoser==2.0.3
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.1
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.8
